test(next): use fixtures and snapshots
