+++ packages/next/src/generators/next-auth/files/pages/api/[...nextauth].ts__template__
