+++ packages/next/src/generators/next-auth/test/fixtures/index.ts
@@ -1,0 +1,4 @@
+export { default as nextAppWithProviders } from './next-app-with-providers';
+export { default as nextAppWithDestructuredProperties } from './next-app-with-destructured-properties';
+export { default as nextAuthEmpty } from './next-auth-empty';
+export { default as nextAuthWithGithub } from './next-auth-with-github';
