+++ packages/next/src/generators/next-auth/test/fixtures/next-app-with-destructured-properties.ts
@@ -1,0 +1,12 @@
+export default `function CustomApp({
+    Component,
+    pageProps: { something, ...pageProperties },
+}: AppProps) {
+    return (
+        <SomeProvider>
+            <Component {...pageProperties} />
+        </SomeProvider>
+    );
+}
+export default CustomApp;
+`;
