+++ packages/next/src/generators/next-auth/test/fixtures/next-app-with-providers.ts
@@ -1,0 +1,15 @@
+export default `function SomeProvider({ children }: PropsWithChildren) {
+    return <>{children}</>;
+}
+function CustomApp({
+    Component,
+    pageProps,
+}: AppProps) {
+    return (
+        <SomeProvider>
+            <Component {...pageProps} />
+        </SomeProvider>
+    );
+}
+export default CustomApp;
+`;
