+++ packages/next/src/generators/next-auth/test/fixtures/next-auth-empty.ts
@@ -1,0 +1,4 @@
+export default `import NextAuth from 'next-auth';
+const nextAuth = NextAuth({});
+default export nextAuth;
+`;
