+++ packages/next/src/generators/next-auth/test/fixtures/next-auth-with-github.ts
@@ -1,0 +1,8 @@
+export default `import NextAuth from 'next-auth';
+const nextAuth = NextAuth({
+  providers: [
+    GithubProvider({})
+  ]
+});
+default export nextAuth;
+`;
